refactor(batchUpdate): move batchUpdate to own file
