+++ src/app/redux/store.ts
@@ -7,35 +7,12 @@ import thunk from 'redux-thunk';
 import * as appConfig from '../../../config/main';
 import { batchUpdate } from './batchUpdate';
 import { IStore } from './IStore';
-import { autoRehydrate } from 'redux-persist';
-const createLogger = require('redux-logger');
+import rootReducer from './reducers';
 
-import { CACHE } from 'actions';
-const { DONE } = CACHE;
+export function configureStore(history: History) {
+  type Enhancer = StoreEnhancerStoreCreator<IStore>;
 
-const doneType = DONE.toString();
-
-const batchUpdate = ({ dispatch }) => next => {
-  let pool = [];
-  const flush = () => {
-    if (pool.length === 0) return;
-    const action = { type: doneType, payload: [...pool] };
-    pool = [];
-    return dispatch(action);
-  };
-  setInterval(flush, 300);
-  return action => {
-    if (action.type === doneType && typeof action.payload === 'number') {
-      pool.push(action.payload);
-      return;
-    }
-
-    return next(action);
-  };
-};
-export function configureStore(history, initialState?: IStore): Redux.Store<IStore> {
-
-  const middlewares: Redux.Middleware[] = [
+  const middlewares: Middleware[] = [
     routerMiddleware(history),
     batchUpdate,
     thunk,
